# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Postal API
  version: "1.0.0"
  description: >
    Postal is a free, no-paywall social-media scheduling & publishing platform.

    This is the client-agnostic JSON API (`/api/v1`) consumed by both web and
    mobile clients. The first supported social platform is X/Twitter.


    ## Envelopes

    Every JSON success response is wrapped in `{ "data": <payload> }` (a `meta`
    object may accompany it). Every error response is `{ "error": { code,
    message, fields?, request_id? } }`. Two endpoints stream raw bodies instead
    of JSON: media download (binary) and analytics CSV export.


    ## Authentication

    Two interchangeable schemes:

    - **Bearer JWT** — `Authorization: Bearer <access_token>` (mobile/API).
    - **Cookie session** — `postal_access` (httpOnly) carries the JWT for web;
      the `postal_csrf` cookie + `X-CSRF-Token` header form a double-submit CSRF
      guard required on state-changing cookie-flow requests. Bearer-only clients
      are CSRF-exempt.


    ## Rate limits & quotas

    Auth endpoints are per-IP throttled; the whole authenticated API has a
    per-user catch-all bucket. Rejections return `429` with `Retry-After` and
    `X-RateLimit-*` headers. Workspaces have durable quotas (channels, pending
    scheduled jobs, media storage) returned as `400` validation errors.
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.postal.example
...
security:
  - bearerAuth: []
  - cookieAuth: []
    csrfToken: []
paths:
...
  /api/v1/workspaces/{workspaceID}/billing/wallet:
    parameters:
      - $ref: "#/components/parameters/workspaceID"
    get:
      tags: [Billing]
      summary: Wallet balance + per-platform publish costs (capability; read)
      responses:
        "200":
          description: Wallet
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/Wallet" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/v1/workspaces/{workspaceID}/billing/ledger:
    parameters:
      - $ref: "#/components/parameters/workspaceID"
    get:
      tags: [Billing]
      summary: Wallet movement history, newest first (capability; read)
      responses:
        "200":
          description: Ledger entries
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/LedgerEntry" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/v1/workspaces/{workspaceID}/billing/topup:
    parameters:
      - $ref: "#/components/parameters/workspaceID"
    post:
      tags: [Billing]
      summary: Start a wallet top-up checkout (capability; manage_workspace)
      description: The charge amount is derived server-side from credits. Returns the provider's hosted checkout URL; crediting happens via the signed webhook, never on redirect.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [provider, credits]
              properties:
                provider: { type: string, enum: [stripe, paystack, dev], description: dev is development-only }
                credits: { type: integer, format: int64, minimum: 1 }
      responses:
        "200":
          description: Hosted checkout URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    required: [checkout_url]
                    properties:
                      checkout_url: { type: string, format: uri }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/v1/billing/webhooks/stripe:
    post:
      tags: [Billing]
      summary: Stripe webhook (public; Stripe-Signature is the auth)
      security: []
      responses:
        "200": { description: Acknowledged (credited, duplicate, or ignored) }
        "401": { description: Signature verification failed }
  /api/v1/billing/webhooks/paystack:
    post:
      tags: [Billing]
      summary: Paystack webhook (public; x-paystack-signature is the auth)
      security: []
      responses:
        "200": { description: Acknowledged (credited, duplicate, or ignored) }
        "401": { description: Signature verification failed }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Authorization: Bearer <access_token>"
    cookieAuth:
      type: apiKey
      in: cookie
      name: postal_access
      description: httpOnly session cookie carrying the access JWT (web flow)
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: Double-submit CSRF token (must equal the postal_csrf cookie) on cookie-flow mutations
  parameters:
    workspaceID:
      name: workspaceID
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Authenticated but lacking the required capability or not a member
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    ValidationError:
      description: Malformed or invalid input
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string, example: invalid_token }
            message: { type: string }
            request_id: { type: string }
            fields:
              type: array
              items:
                type: object
                properties:
                  field: { type: string }
                  message: { type: string }
      required: [error]
...
    Wallet:
      type: object
      required: [workspace_id, balance, publish_costs, updated_at]
      properties:
        workspace_id: { type: string, format: uuid }
        balance: { type: integer, format: int64, description: credits }
        publish_costs:
          type: object
          additionalProperties: { type: integer, format: int64 }
          description: per-publish cost in credits by platform; absent platforms are free
        updated_at: { type: string, format: date-time }
    LedgerEntry:
      type: object
      required: [id, workspace_id, kind, credits, reference, note, created_at]
      properties:
        id: { type: string, format: uuid }
        workspace_id: { type: string, format: uuid }
        kind: { type: string, enum: [topup, publish_charge, refund, adjustment] }
        credits: { type: integer, format: int64, description: signed; charges are negative }
        reference: { type: string }
        note: { type: string }
        created_at: { type: string, format: date-time }