# internal/platform/db/sqlc/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.27.0

package sqlc

import (
	"github.com/google/uuid"
	"github.com/jackc/pgx/v5/pgtype"
)

type AuditLog struct {
	ID          int64              `json:"id"`
	WorkspaceID *uuid.UUID         `json:"workspace_id"`
	ActorUserID *uuid.UUID         `json:"actor_user_id"`
	Action      string             `json:"action"`
	Target      string             `json:"target"`
	Metadata    []byte             `json:"metadata"`
	Ip          string             `json:"ip"`
	CreatedAt   pgtype.Timestamptz `json:"created_at"`
}

type Channel struct {
	ID                uuid.UUID          `json:"id"`
	WorkspaceID       uuid.UUID          `json:"workspace_id"`
	Platform          string             `json:"platform"`
	PlatformAccountID string             `json:"platform_account_id"`
	Handle            string             `json:"handle"`
	DisplayName       string             `json:"display_name"`
	Status            string             `json:"status"`
	ConnectedBy       *uuid.UUID         `json:"connected_by"`
	CreatedAt         pgtype.Timestamptz `json:"created_at"`
	UpdatedAt         pgtype.Timestamptz `json:"updated_at"`
}

type ChannelCredential struct {
	ChannelID             uuid.UUID          `json:"channel_id"`
	EncryptedAccessToken  []byte             `json:"encrypted_access_token"`
	EncryptedRefreshToken []byte             `json:"encrypted_refresh_token"`
	Scopes                []string           `json:"scopes"`
	ExpiresAt             pgtype.Timestamptz `json:"expires_at"`
	KeyVersion            int32              `json:"key_version"`
	CreatedAt             pgtype.Timestamptz `json:"created_at"`
	UpdatedAt             pgtype.Timestamptz `json:"updated_at"`
}

type EmailVerificationToken struct {
	ID         uuid.UUID          `json:"id"`
	UserID     uuid.UUID          `json:"user_id"`
	TokenHash  string             `json:"token_hash"`
	ExpiresAt  pgtype.Timestamptz `json:"expires_at"`
	ConsumedAt pgtype.Timestamptz `json:"consumed_at"`
	CreatedAt  pgtype.Timestamptz `json:"created_at"`
}

type MediaAsset struct {
	ID          uuid.UUID          `json:"id"`
	WorkspaceID uuid.UUID          `json:"workspace_id"`
	Kind        string             `json:"kind"`
	StorageKey  string             `json:"storage_key"`
	Mime        string             `json:"mime"`
	Width       int32              `json:"width"`
	Height      int32              `json:"height"`
	DurationMs  int32              `json:"duration_ms"`
	Bytes       int64              `json:"bytes"`
	Status      string             `json:"status"`
	CreatedAt   pgtype.Timestamptz `json:"created_at"`
}

type MetricPollState struct {
	ChannelID      uuid.UUID          `json:"channel_id"`
	PlatformPostID string             `json:"platform_post_id"`
	LastPolledAt   pgtype.Timestamptz `json:"last_polled_at"`
	Done           bool               `json:"done"`
}

type MetricSnapshot struct {
	ID             int64              `json:"id"`
	WorkspaceID    uuid.UUID          `json:"workspace_id"`
	ChannelID      uuid.UUID          `json:"channel_id"`
	PostID         *uuid.UUID         `json:"post_id"`
	PlatformPostID string             `json:"platform_post_id"`
	Metric         string             `json:"metric"`
	Value          int64              `json:"value"`
	CapturedAt     pgtype.Timestamptz `json:"captured_at"`
}

type PasswordResetToken struct {
	ID         uuid.UUID          `json:"id"`
	UserID     uuid.UUID          `json:"user_id"`
	TokenHash  string             `json:"token_hash"`
	ExpiresAt  pgtype.Timestamptz `json:"expires_at"`
	ConsumedAt pgtype.Timestamptz `json:"consumed_at"`
	CreatedAt  pgtype.Timestamptz `json:"created_at"`
}

type Post struct {
	ID           uuid.UUID          `json:"id"`
	WorkspaceID  uuid.UUID          `json:"workspace_id"`
	AuthorUserID *uuid.UUID         `json:"author_user_id"`
	Status       string             `json:"status"`
	CreatedAt    pgtype.Timestamptz `json:"created_at"`
	UpdatedAt    pgtype.Timestamptz `json:"updated_at"`
}

type PostVariant struct {
	ID              uuid.UUID          `json:"id"`
	PostID          uuid.UUID          `json:"post_id"`
	ChannelID       uuid.UUID          `json:"channel_id"`
	Body            string             `json:"body"`
	MediaRefs       []byte             `json:"media_refs"`
	PlatformOptions []byte             `json:"platform_options"`
	CreatedAt       pgtype.Timestamptz `json:"created_at"`
	UpdatedAt       pgtype.Timestamptz `json:"updated_at"`
}

type PublishResult struct {
	ID             int64              `json:"id"`
	ChannelID      uuid.UUID          `json:"channel_id"`
	PostID         *uuid.UUID         `json:"post_id"`
	IdempotencyKey string             `json:"idempotency_key"`
	PlatformPostID string             `json:"platform_post_id"`
	RawResponse    []byte             `json:"raw_response"`
	PublishedAt    pgtype.Timestamptz `json:"published_at"`
}

type ScheduleSlot struct {
	ID        uuid.UUID          `json:"id"`
	ChannelID uuid.UUID          `json:"channel_id"`
	DayOfWeek int16              `json:"day_of_week"`
	TimeOfDay string             `json:"time_of_day"`
	Timezone  string             `json:"timezone"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
}

type ScheduledJob struct {
	ID          uuid.UUID          `json:"id"`
	PostID      uuid.UUID          `json:"post_id"`
	ChannelID   uuid.UUID          `json:"channel_id"`
	RunAt       pgtype.Timestamptz `json:"run_at"`
	Status      string             `json:"status"`
	Attempts    int32              `json:"attempts"`
	LastError   string             `json:"last_error"`
	AsynqTaskID string             `json:"asynq_task_id"`
	CreatedAt   pgtype.Timestamptz `json:"created_at"`
	UpdatedAt   pgtype.Timestamptz `json:"updated_at"`
}

type SchemaSmoke struct {
	ID        int64              `json:"id"`
	Note      string             `json:"note"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
}

type User struct {
	ID            uuid.UUID          `json:"id"`
	Email         string             `json:"email"`
	PasswordHash  string             `json:"password_hash"`
	EmailVerified bool               `json:"email_verified"`
	Status        string             `json:"status"`
	CreatedAt     pgtype.Timestamptz `json:"created_at"`
	UpdatedAt     pgtype.Timestamptz `json:"updated_at"`
}

type Wallet struct {
	WorkspaceID uuid.UUID          `json:"workspace_id"`
	Balance     int64              `json:"balance"`
	UpdatedAt   pgtype.Timestamptz `json:"updated_at"`
}

type WalletLedger struct {
	ID          uuid.UUID          `json:"id"`
	WorkspaceID uuid.UUID          `json:"workspace_id"`
	Kind        string             `json:"kind"`
	Credits     int64              `json:"credits"`
	Reference   string             `json:"reference"`
	Note        string             `json:"note"`
	CreatedAt   pgtype.Timestamptz `json:"created_at"`
}

type Workspace struct {
	ID          uuid.UUID          `json:"id"`
	Name        string             `json:"name"`
	OwnerUserID uuid.UUID          `json:"owner_user_id"`
	Plan        string             `json:"plan"`
	CreatedAt   pgtype.Timestamptz `json:"created_at"`
}

type WorkspaceMember struct {
	WorkspaceID uuid.UUID          `json:"workspace_id"`
	UserID      uuid.UUID          `json:"user_id"`
	Role        string             `json:"role"`
	Permissions []string           `json:"permissions"`
	CreatedAt   pgtype.Timestamptz `json:"created_at"`
}